Init: Migration from adiwg/mdKeywords and New Validation (#3)
* feat: Add boilerplate for GNIS Harvester & Initial Migration from mdKeywords

This commit adds the initial boilerplate code for the GNIS Harvester module and includes the migration of content from the adiwg/mdKeywords repository. It also includes the necessary configuration files and scripts for the harvesters.

* feat: Add GNIS Harvester module and migrate content from mdKeywords

This commit adds the initial boilerplate code for the GNIS Harvester module and includes the migration of content from the adiwg/mdKeywords repository. It also includes the necessary configuration files and scripts for the harvesters.

* chore: Update .gitignore

* Remove 'broader' as a key in the keyword schema

* Add dependencies for babel and other dev tools.

Converting to ES6, refactoring as needed.

Removing husky and pre-commit hooks.

* Add build script and supporting. Update yarn.lock. Add rimraf.

* Converted to ES6+ modules

* chore: Update .gitignore

* feat: Add GNIS Harvester module and migrate content from mdKeywords

* chore: Update start, gcmd, gnis, sciencebase, and usgs scripts to include yarn build command

This commit updates the start, gcmd, gnis, sciencebase, and usgs scripts in the package.json file to include the yarn build command before running the respective harvesters. This ensures that the code is built before executing the harvesters, improving the reliability and consistency of the harvesting process.

* feat: Update GCMD harvester to fetch and store thesaurus and keywords

* feat: Add GCMD thesaurus and keyword files

This commit adds the GCMD thesaurus and keyword files to the resources directory. These files provide metadata language and organization personnel role information. They are essential for the GCMD harvester module to fetch and store thesaurus and keyword data.

* feat: Add GCMD thesaurus and keyword files

* feat: Update GCMD harvester to fetch and store thesaurus and keywords

* feat: Add ScienceBase thesaurus and keyword files

* Refactor thesaurus JSON files

This commit refactors the thesaurus JSON files in the resources/thesaurus directory. The "description" field in each file was updated from "No description available." to an empty string. This change was made to improve the clarity and consistency of the thesaurus data.

* Refactor thesaurus JSON files

* feat: Update eslint.config.js with import plugin

* Refactor build to use tests directory instead of scripts

* Refactor USGS harvester to generate and store thesaurus and keyword files

This commit refactors the USGS harvester to generate and store thesaurus and keyword files. The harvester now fetches the thesaurus data from the USGS website and generates the keywords based on the retrieved data. The thesaurus configuration file is updated to include the keywords URL. The generated files are stored in the appropriate directories. This change improves the functionality and data consistency of the USGS harvester.

* Add USGS thesaurus configuration and keyword files

* Add buildManifest script and add it to the package.json scripts.

* refactor: update label field in sciencebase.js to include metadata.name if available

* Refactor thesaurus JSON files

* Add manifest file

* Refactor thesaurus JSON files

This commit refactors the thesaurus JSON files in the resources/thesaurus directory. The "description" field in each file was updated from "No description available." to an empty string. This change was made to improve the clarity and consistency of the thesaurus data.

* Add copySchemas script

* chore: Add @babel/plugin-syntax-import-attributes to .babelrc

* build: Update build script to include copying schemas and generating tests

* refactor: Remove parentId key if null in gcmd.js

* Refactor import statements in schemas/index.js

* Refactor keyword schema to use array of keywords instead of object

* Refactor JSON schema file IDs to use "$id" instead of "id"

* Refactor tests to validate keywords and thesaurus files

This commit refactors the test files for keywords and thesaurus to improve the validation process. The tests now iterate over the files in the respective directories and validate each file individually. This change ensures that all keywords and thesaurus files are properly validated, improving the reliability of the application.

Refactor keywords.test.js and thesauri.test.js

* Update resource file

* Update resource file(s)

* Update resource file(s)

* Update README.md

* refactor: Remove parentId key if null in gcmd.js

* Update resource files

* Update resource files