chore: Update manifest.json schema to support an array of objects
The commit modifies the manifest.json schema to support an array of objects instead of a single object. Each object in the array must have a "name" (string) and a "url" (string) property. This change allows for multiple thesaurus configurations to be defined in the manifest.json file. The existing properties and their descriptions remain the same.