Fix/datetime (#4)
* Add dayjs, increment patch version

* feat: Fix the date format in thesaurus citation. Remove keywordType and keywords keys. Add keywordUrl to gcmd configs.

* refactor: Remove keywordType and keywords keys. Add keywordsUrl.

* feat: Update thesaurus citation date format and import dayjs library

* feat: Update thesaurusConfig.json and usgs.js

The commit updates the thesaurusConfig.json file by removing the "date" and "dateType" keys and replacing them with empty values. It also removes the "keywordType" and "keywords" keys. Additionally, the usgs.js file imports the dayjs library and adds a new function to generate thesaurus configuration with an updated "date" value.

* feat: Remove unnecessary keys from thesaurusConfig.json

The commit removes the "date" and "dateType" keys from the thesaurusConfig.json file and replaces them with empty values. It also removes the "keywordType" and "keywords" keys.

* feat: Update usgs thesaurus config

* feat: Update ScienceBase Thesaurus Configs

* chore: Update ScienceBase Thesaurus Configs

* chore: Update GCMD Thesaurus Configs

* chore: Update GCMD Thesaurus Configs

* chore: Remove harvest-keywords.yml github action. Add test-resoureces.yml github action.

* chore: Update Node.js version to 20 in test-resources.yml

* chore: Refactor copyGnisData script to handle non-existent source directory

The commit refactors the copyGnisData script to handle the case when the source directory "./harvesters/gnis/data" does not exist. It adds a try-catch block to check if the directory exists using fs.pathExists and skips the copy operation if it doesn't. If the directory exists, it copies the data to the destination directory "./dist/harvesters/gnis/data" and logs a success message. If an error occurs during the copy operation, it logs the error message.

* chore: Update manifest.json schema to support an array of objects

The commit modifies the manifest.json schema to support an array of objects instead of a single object. Each object in the array must have a "name" (string) and a "url" (string) property. This change allows for multiple thesaurus configurations to be defined in the manifest.json file. The existing properties and their descriptions remain the same.

* chore: Update GCMD and ScienceBase Thesaurus Configs