Features/10 develop gnis harvester (#13)
* feat: Add `readline` and `xml2js` dependencies

The commit adds the `readline` and `xml2js` dependencies to the `package.json` file. This is necessary for the project to use the `readline` module for reading input from the command line and the `xml2js` module for parsing XML data. These dependencies will enhance the functionality of the project and enable it to handle XML data effectively.

* feat: Add GNIS harvester for Antarctic

* feat: Refactor filestream variable in parseTxtFileToJson function

The commit refactors the `filestream` variable in the `parseTxtFileToJson` function in `gnisUtils.js`. It updates the variable assignment to include the `fullRelativePath` for better readability and clarity. This change ensures that the correct file path is used when creating the read stream for parsing the TXT file to JSON.

* feat: Refactor GNIS harvester for Antarctica

The commit refactors the GNIS harvester for Antarctica in the `antarctica.js` file. It updates the import statements to remove unnecessary file extensions and removes unused variables and imports. This change improves the readability and maintainability of the code.

* feat: Working on Domestic Names

* feat: Refactor DomesticNames to optimize processing the data. Also add additional information to output files.

* chore: Add uuid

* feat: Domestic Names

* chore: Add new resource files. Updated manifest.

* feat: Thesaurus Config for Domestic Names.

* chore: Remove keyword files added in error.

* chore: Refactor feature class id, removing generated uuid in place of consistent static value.

* chore: Refactor feature class id to remove spaces and symbols to be replaced with hyphens and convert all characters to lowercase.

* chore: Update resource files

* chore: Refactor state uuid to replace spaces and symbols with hyphens and convert all characters to lowercase.

* chore: Update resource files

* chore: Refactor filename to convert symbols and spaces to hyphens and convert all characters to lowercase.

* chore: Update manifest files.

* chore: Update resource files.

* chore: Update actions/checkout to v3 in test-resources.yml

* chore: Remove files - fixing a casing error. GitHub really needs to figure out how to be case sensitive.

* chore: Remove manifest files. Fixing errors.

* chore: Update resource files.

* chore: Move extra scripts and provide short explanation about what they are.