Sync dev -> main (#20)
* Features/10 develop gnis harvester (#13)

* feat: Add `readline` and `xml2js` dependencies

The commit adds the `readline` and `xml2js` dependencies to the `package.json` file. This is necessary for the project to use the `readline` module for reading input from the command line and the `xml2js` module for parsing XML data. These dependencies will enhance the functionality of the project and enable it to handle XML data effectively.

* feat: Add GNIS harvester for Antarctic

* feat: Refactor filestream variable in parseTxtFileToJson function

The commit refactors the `filestream` variable in the `parseTxtFileToJson` function in `gnisUtils.js`. It updates the variable assignment to include the `fullRelativePath` for better readability and clarity. This change ensures that the correct file path is used when creating the read stream for parsing the TXT file to JSON.

* feat: Refactor GNIS harvester for Antarctica

The commit refactors the GNIS harvester for Antarctica in the `antarctica.js` file. It updates the import statements to remove unnecessary file extensions and removes unused variables and imports. This change improves the readability and maintainability of the code.

* feat: Working on Domestic Names

* feat: Refactor DomesticNames to optimize processing the data. Also add additional information to output files.

* chore: Add uuid

* feat: Domestic Names

* chore: Add new resource files. Updated manifest.

* feat: Thesaurus Config for Domestic Names.

* chore: Remove keyword files added in error.

* chore: Refactor feature class id, removing generated uuid in place of consistent static value.

* chore: Refactor feature class id to remove spaces and symbols to be replaced with hyphens and convert all characters to lowercase.

* chore: Update resource files

* chore: Refactor state uuid to replace spaces and symbols with hyphens and convert all characters to lowercase.

* chore: Update resource files

* chore: Refactor filename to convert symbols and spaces to hyphens and convert all characters to lowercase.

* chore: Update manifest files.

* chore: Update resource files.

* chore: Update actions/checkout to v3 in test-resources.yml

* chore: Remove files - fixing a casing error. GitHub really needs to figure out how to be case sensitive.

* chore: Remove manifest files. Fixing errors.

* chore: Update resource files.

* chore: Move extra scripts and provide short explanation about what they are.

* Update reference to dev branch (#14)

* feat: Add GNIS Antarctica and GNIS Domestic California thesaurus files (#15)

The commit adds the GNIS Antarctica and GNIS Domestic California thesaurus files to the repository. These files contain geographic nomenclature data for Antarctica and domestic locations in California, respectively. They are part of the Geographic Names Information System (GNIS) and serve as the official repository of geographic names data for the Federal Government. These additions enhance the repository's resources for geographic naming and provide valuable data for various applications.

* Features/11 develop nalt harvester (#16)

* feat: Add boilerplate for NALT harvester.

* feat: NALT Harvester Functional.

Mostly done, needs minor tweaks to finish it up. Stil needs to produce thesaurus config file.

* feat: Complete NALT harvester.

* chore: Add new NALT artifacts.

* Features/12-refactor-lithography-harvester (#17)

* refactor: Cleaned up and improved the harvester code. Added error handling and retries when request for term fails. Saving progress.

* refactor: Generalizing the USGS (formerly the lithography) harvester. The former USGS Thesaurus harvester has been renamed and kept as a backup, but is now deprecated. The USGS harvester can produce a thesaurus for all the vocabularies including the USGS Thesaurus itself.

Note: Running the USGS harvester takes hours and it should probably be refactored to handle individual vocabularies instead of all of them at once.

* chore: Run USGS harvester. Saving artifacts from harvest operation. Updating manifest files.

* fix: Output format fixed and files updated.

* Chore/update readme (#19)

* docs: Update README.md with detailed structure and resource descriptions

* refactor: Improve performance of data processing functions